Change IndexingService and remove OnePageService

+++ src/main/java/searchengine/model/Site.java
@@ -5,7 +5,6 @@ import lombok.Setter;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Getter
 @Setter
@@ -14,7 +13,7 @@ public class Site {
 
     @Id
     @GeneratedValue (strategy = GenerationType.IDENTITY)
-    private Integer Id;
+    private Long Id;
 
     @Enumerated(EnumType.STRING)
     @Column(columnDefinition = "ENUM('INDEXING', 'INDEXED', 'FAILED')" , nullable = false)
